+++ samples/docsample.cpp
@@ -1,4 +1,4 @@
-#include "../source/utf8.h"
+#include "utf8cpp/utf8.h"
 #include <iostream>
 #include <fstream>
 #include <string>
