+++ tests/negative.cpp
@@ -1,4 +1,4 @@
-#include "utf8.h"
+#include "utf8cpp/utf8.h"
 using namespace utf8;
 
 #include <string>
