+++ tests/test_checked_api.cpp
@@ -1,5 +1,5 @@
 #include "gtest/gtest.h"
-#include "utf8.h"
+#include "utf8cpp/utf8.h"
 
 #include <string>
 #include <vector>
