+++ tests/test_checked_iterator.cpp
@@ -1,5 +1,5 @@
 #include "gtest/gtest.h"
-#include "utf8.h"
+#include "utf8cpp/utf8.h"
 
 using namespace utf8;
