+++ tests/test_cpp11.cpp
@@ -1,5 +1,5 @@
 #include "gtest/gtest.h"
-#include "utf8.h"
+#include "utf8cpp/utf8.h"
 #include <string>
 using namespace utf8;
 using namespace std;
