+++ tests/test_unchecked_api.cpp
@@ -1,5 +1,5 @@
 #include "gtest/gtest.h"
-#include "utf8/unchecked.h"
+#include "utf8cpp/utf8/unchecked.h"
 
 #include <string>
 #include <vector>
